chore: move to its own package
