+++ packages/mask/src/plugin-infra/register.ts
@@ -31,7 +31,7 @@ import '../plugins/MaskBox'
 import '../plugins/NextID'
 // import '../plugins/Profile'
 import '../plugins/Pets'
-import '../plugins/CyberConnect'
+import '@masknet/plugin-cyber-connect'
 // import '../plugins/NFT'
 // import '../plugins/Airdrop'
 import '../plugins/CryptoartAI'
