+++ packages/plugins/CyberConnect/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@masknet/plugin-cyber-connect",
+  "private": true,
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "dependencies": {
+    "@masknet/plugin-infra": "workspace:*",
+    "@masknet/theme": "workspace:*",
+    "@masknet/web3-shared-evm": "workspace:*",
+    "@masknet/shared-base": "workspace:*",
+    "@cyberlab/cyberconnect": "^4.2.0",
+    "boring-avatars": "^1.5.8",
+    "react-use": "^17.3.2"
+  }
+}
