+++ packages/plugins/CyberConnect/src/assets/CyberConnectIcon.tsx
