+++ packages/plugins/CyberConnect/src/assets/Foundation.png
