+++ packages/plugins/CyberConnect/src/assets/Opensea.png
