+++ packages/plugins/CyberConnect/src/env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@masknet/global-types/webpack" />
