+++ packages/plugins/tsconfig.json
@@ -7,6 +7,7 @@
     { "path": "./Wallet/" },
     { "path": "./FileService/" },
     { "path": "./RSS3/" },
-    { "path": "./DAO/" }
+    { "path": "./DAO/" },
+    { "path": "./CyberConnect/" }
   ]
 }
