+++ packages/shared-base/src/utils/index.ts
@@ -9,3 +9,4 @@ export * from './pollingTask'
 export * from './sessionStorageCache'
 export * from './subscription'
 export * from './abortSignal'
+export * from './getAssetAsBlobURL'
