+++ pnpm-workspace.yaml
@@ -33,6 +33,7 @@ packages:
   - 'packages/plugins/DAO'
   - 'packages/plugins/RSS3'
   - 'packages/plugins/Solana'
+  - 'packages/plugins/CyberConnect'
   - 'packages/external-plugin-previewer'
   - 'packages/backup-format'
   - 'packages/encryption'
